Refactor main.rs: Split into modules and remove test-specific code
- Extract CLI argument parsing to src/cli/args.rs
- Extract helper functions to src/cli/helpers.rs
- Remove Mock Test Script detection logic
- Remove hardcoded test references (mock.sh, sonesuke@Air)
- Simplify rule processing by removing script state tracking

This reduces main.rs by ~150 lines and improves code organization.
