+++ src/main.rs
@@ -1,60 +1,23 @@
 mod agent;
+mod cli;
 mod queue;
 mod ruler;
 mod workflow;
 
 use anyhow::{Context, Result};
-use clap::{Args, Parser, Subcommand};
+use clap::Parser;
+use cli::{
+    execute_entry_action, execute_periodic_entry, resolve_entry_task_placeholders, Cli, Commands,
+};
 use queue::{create_shared_manager, QueueExecutor};
 use ruler::decision::decide_action;
 use ruler::entry::TriggerType;
-use ruler::rule::resolve_task_placeholder_in_vec;
 use ruler::Ruler;
 use std::path::PathBuf;
 use std::sync::Arc;
 use tokio::signal;
 use tokio::time::interval;
 use tracing::info;
-
-#[derive(Parser, Debug)]
-#[command(author, version, about, long_about = None)]
-struct Cli {
-    #[command(subcommand)]
-    command: Option<Commands>,
-
-    /// Path to config YAML file
-    #[arg(short, long, global = true)]
-    rules: Option<PathBuf>,
-
-    /// Enable debug logging for internal details
-    #[arg(short, long, global = true)]
-    debug: bool,
-}
-
-#[derive(Subcommand, Debug)]
-enum Commands {
-    /// Load and display rules
-    Show(ShowArgs),
-    /// Test rule matching against capture text
-    Test(TestArgs),
-}
-
-#[derive(Args, Debug)]
-struct ShowArgs {
-    /// Path to config YAML file
-    #[arg(short, long, default_value = "config.yaml")]
-    rules: PathBuf,
-}
-
-#[derive(Args, Debug)]
-struct TestArgs {
-    /// Path to config YAML file
-    #[arg(short, long, default_value = "config.yaml")]
-    rules: PathBuf,
-    /// Capture text to test against rules
-    #[arg(short, long)]
-    capture: String,
-}
 
 #[tokio::main]
 async fn main() -> Result<()> {
@@ -91,7 +54,6 @@ async fn main() -> Result<()> {
             println!("🎯 RuleAgents started");
             println!("📂 Config file: {}", rules_path.display());
             println!("🌐 Terminal available at: http://localhost:{}", base_port);
-            println!("💡 Type 'entry' in the terminal to start mock.sh");
             println!("🛑 Press Ctrl+C to stop");
 
             // Create agent directly
@@ -195,7 +157,6 @@ async fn main() -> Result<()> {
             }
 
             let mut last_output_lines: Vec<String> = Vec::new();
-            let mut is_script_running = false;
 
             loop {
                 tokio::select! {
@@ -276,23 +237,10 @@ async fn main() -> Result<()> {
                                     println!("📄 Initial buffer content: {:?}", &clean_content[..clean_content.len().min(200)]);
                                 }
 
-                                // === SCRIPT STATE DETECTION ===
-                                // Monitor script lifecycle by detecting key patterns in new content
-                                if new_content.contains("=== Mock Test Script ===") {
-                                    is_script_running = true;
-                                    println!("🎬 Script started");
-                                } else if is_script_running && (new_content.contains("MISSION COMPLETE") || new_content.contains("operation has been successfully completed")) {
-                                    is_script_running = false;
-                                    println!("💤 Script completed - Agent returned to idle state");
-                                }
 
                                 // === RULE PROCESSING ON NEW CONTENT ===
                                 // Apply rules only to the newly detected content
-                                if !new_content.is_empty() && is_script_running && new_content.contains("Do you want to proceed") && !new_content.contains("MISSION COMPLETE") {
-                                    println!("🎯 Found 'Do you want to proceed' in new content!");
-                                    println!("🔍 New content: {:?}", &new_content[..new_content.len().min(300)]);
-                                    println!("🔍 Script running: {}", is_script_running);
-                                    println!("🔍 Contains MISSION COMPLETE: {}", new_content.contains("MISSION COMPLETE"));
+                                if !new_content.is_empty() {
 
                                     let action = ruler.decide_action_for_capture(&new_content).await;
                                     match action {
@@ -348,10 +296,6 @@ async fn main() -> Result<()> {
                                     }
                                 }
 
-                                // Check for idle state in new content
-                                if !new_content.is_empty() && !is_script_running && new_content.contains("sonesuke@Air") && new_content.contains("%") {
-                                    println!("⏸️ Idle state detected in new content");
-                                }
 
                                 // Update the stored output for next comparison
                                 // Store as single string since HT terminal sends everything as one buffer
@@ -413,152 +357,4 @@ async fn main() -> Result<()> {
     }
 
     Ok(())
-}
-
-/// Execute a periodic entry action (with agent context)
-async fn execute_periodic_entry(
-    entry: &ruler::entry::CompiledEntry,
-    queue_manager: &queue::SharedQueueManager,
-    agent: Option<&agent::Agent>,
-) -> Result<()> {
-    match &entry.action {
-        ruler::types::ActionType::SendKeys(keys) => {
-            if let Some(agent) = agent {
-                println!(
-                    "🤖 Executing periodic entry '{}' → Sending: {:?}",
-                    entry.name, keys
-                );
-                for key in keys {
-                    agent.send_keys(key).await?;
-                }
-            } else {
-                println!(
-                    "⚠️ Periodic entry '{}' has SendKeys action - skipping (no agent context)",
-                    entry.name
-                );
-            }
-        }
-        ruler::types::ActionType::Workflow(workflow_name, args) => {
-            println!(
-                "🔄 Executing periodic entry '{}' → Workflow: {} {:?}",
-                entry.name, workflow_name, args
-            );
-            // TODO: Implement workflow execution
-        }
-        ruler::types::ActionType::Enqueue { queue, command } => {
-            println!(
-                "📦 Executing periodic entry '{}' → Enqueue to '{}': {}",
-                entry.name, queue, command
-            );
-            let executor = QueueExecutor::new(queue_manager.clone());
-            let count = executor.execute_and_enqueue(queue, command).await?;
-            println!("✅ Enqueued {} items to queue '{}'", count, queue);
-        }
-        ruler::types::ActionType::EnqueueDedupe { queue, command } => {
-            println!(
-                "📦 Executing periodic entry '{}' → EnqueueDedupe to '{}': {}",
-                entry.name, queue, command
-            );
-            let executor = QueueExecutor::new(queue_manager.clone());
-            let count = executor.execute_and_enqueue_dedupe(queue, command).await?;
-            println!(
-                "✅ Enqueued {} new items to dedupe queue '{}'",
-                count, queue
-            );
-        }
-    }
-    Ok(())
-}
-
-/// Execute an entry action using the appropriate mechanism
-async fn execute_entry_action(
-    agent: &agent::Agent,
-    entry: &ruler::entry::CompiledEntry,
-    queue_manager: &queue::SharedQueueManager,
-) -> Result<()> {
-    match &entry.action {
-        ruler::types::ActionType::SendKeys(keys) => {
-            println!("🤖 Executing entry '{}' → Sending: {:?}", entry.name, keys);
-            for key in keys {
-                if key == "\\r" || key == "\r" {
-                    if let Err(e) = agent.send_keys("\r").await {
-                        eprintln!("❌ Error sending key: {}", e);
-                    }
-                } else if let Err(e) = agent.send_keys(key).await {
-                    eprintln!("❌ Error sending key: {}", e);
-                }
-                tokio::time::sleep(tokio::time::Duration::from_millis(100)).await;
-            }
-        }
-        ruler::types::ActionType::Workflow(workflow_name, args) => {
-            println!(
-                "🔄 Executing entry '{}' → Workflow: {} {:?}",
-                entry.name, workflow_name, args
-            );
-            // TODO: Implement workflow execution
-        }
-        ruler::types::ActionType::Enqueue { queue, command } => {
-            println!(
-                "📦 Executing entry '{}' → Enqueue to '{}': {}",
-                entry.name, queue, command
-            );
-            let executor = QueueExecutor::new(queue_manager.clone());
-            let count = executor.execute_and_enqueue(queue, command).await?;
-            println!("✅ Enqueued {} items to queue '{}'", count, queue);
-        }
-        ruler::types::ActionType::EnqueueDedupe { queue, command } => {
-            println!(
-                "📦 Executing entry '{}' → EnqueueDedupe to '{}': {}",
-                entry.name, queue, command
-            );
-            let executor = QueueExecutor::new(queue_manager.clone());
-            let count = executor.execute_and_enqueue_dedupe(queue, command).await?;
-            println!(
-                "✅ Enqueued {} new items to dedupe queue '{}'",
-                count, queue
-            );
-        }
-    }
-    Ok(())
-}
-
-/// Resolve <task> placeholders in entry action with actual task value
-fn resolve_entry_task_placeholders(
-    entry: &ruler::entry::CompiledEntry,
-    task_value: &str,
-) -> ruler::entry::CompiledEntry {
-    let resolved_action = match &entry.action {
-        ruler::types::ActionType::SendKeys(keys) => {
-            ruler::types::ActionType::SendKeys(resolve_task_placeholder_in_vec(keys, task_value))
-        }
-        ruler::types::ActionType::Workflow(workflow_name, args) => {
-            let resolved_workflow =
-                ruler::rule::resolve_task_placeholder(workflow_name, task_value);
-            let resolved_args = resolve_task_placeholder_in_vec(args, task_value);
-            ruler::types::ActionType::Workflow(resolved_workflow, resolved_args)
-        }
-        ruler::types::ActionType::Enqueue { queue, command } => {
-            let resolved_queue = ruler::rule::resolve_task_placeholder(queue, task_value);
-            let resolved_command = ruler::rule::resolve_task_placeholder(command, task_value);
-            ruler::types::ActionType::Enqueue {
-                queue: resolved_queue,
-                command: resolved_command,
-            }
-        }
-        ruler::types::ActionType::EnqueueDedupe { queue, command } => {
-            let resolved_queue = ruler::rule::resolve_task_placeholder(queue, task_value);
-            let resolved_command = ruler::rule::resolve_task_placeholder(command, task_value);
-            ruler::types::ActionType::EnqueueDedupe {
-                queue: resolved_queue,
-                command: resolved_command,
-            }
-        }
-    };
-
-    ruler::entry::CompiledEntry {
-        name: entry.name.clone(),
-        trigger: entry.trigger.clone(),
-        action: resolved_action,
-        concurrency: entry.concurrency,
-    }
 }
